requirements.txt
flask==2.2.3
google-generativeai==0.3.2
python-dotenv==0.21.1
firebase-admin==6.2.0
requests==2.28.2
python-dateutil==2.8.2
werkzeug==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
gunicorn==20.1.0 
transformers==4.37.2
torch==2.2.0 